requirements.txt
Flask==3.0.0
pdfplumber==0.10.3
python-docx==1.1.0
PyMuPDF==1.23.8
gunicorn==21.2.0